+++ packages/database-mongodb-node/src/con.posts.js
@@ -59,25 +59,38 @@ const get = (driver) => get_regular(driver, col(driver));
  * @returns {db_col["remove"]}
  */
 const remove = (driver) => {
-  return async (id) => {
-    const objid = to_objid(id)
-    // const item = await col(driver).findOne(handle_or_id(id));
+  return async (id_or_handle) => {
+    const item = await col(driver).findOne(handle_or_id(id_or_handle));
+    if(!item) return;
+    const objid = to_objid(item.id)
+    const session = driver.mongo_client.startSession();
 
-    ////
-    // STOREFRONTS --> POSTS RELATION
-    ////
-    await driver.storefronts._col.updateMany(
-      { '_relations.posts.ids' : objid },
-      { 
-        $pull: { '_relations.posts.ids': objid, },
-        $unset: { [`_relations.posts.entries.${objid.toString()}`]: '' },
-      },
-    );
+    try {
+      await session.withTransaction(
+        async () => {
+          ////
+          // STOREFRONTS --> POSTS RELATION
+          ////
+          await driver.storefronts._col.updateMany(
+            { '_relations.posts.ids' : objid },
+            { 
+              $pull: { '_relations.posts.ids': objid, },
+              $unset: { [`_relations.posts.entries.${objid.toString()}`]: '' },
+            },
+            { session }
+          );
 
-    // DELETE ME
-    const res = await col(driver).findOneAndDelete( 
-      { _id: objid }
-    );
+          // DELETE ME
+          const res = await col(driver).findOneAndDelete( 
+            { _id: objid },
+            { session }
+          );
+
+        }
+      );
+    } finally {
+      await session.endSession();
+    }
 
     return
   }
