+++ packages/database-mongodb-node/src/con.shipping.js
@@ -22,30 +22,42 @@ const col = (d) => d.collection('shipping_methods');
 const upsert = (driver) => {
   return async (data) => {
     const objid = to_objid(data.id);
-    const filter = { _id: objid };
     const replacement = { ...data };
-    const options = { upsert: true };
+    const session = driver.mongo_client.startSession();
 
-    ////
-    // STOREFRONTS --> SHIPPING RELATION
-    ////
-    await driver.storefronts._col.updateMany(
-      { '_relations.shipping_methods.ids' : objid },
-      { $set: { [`_relations.shipping_methods.entries.${objid.toString()}`]: data } },
-    );
+    try {
+      await session.withTransaction(
+        async () => {
+          ////
+          // STOREFRONTS --> SHIPPING RELATION
+          ////
+          await driver.storefronts._col.updateMany(
+            { '_relations.shipping_methods.ids' : objid },
+            { $set: { [`_relations.shipping_methods.entries.${objid.toString()}`]: data } },
+            { session }
+          );
 
-    ////
-    // REPORT IMAGES USAGE
-    ////
-    await report_document_media(driver)(data);
+          ////
+          // REPORT IMAGES USAGE
+          ////
+          await report_document_media(driver)(data, session);
 
-    // SAVE ME
-    const res = await col(driver).replaceOne(
-      filter, replacement, options
-    );
+          // SAVE ME
+          const res = await col(driver).replaceOne(
+            { _id: objid }, 
+            replacement, 
+            { session, upsert: true }
+          );
 
+        }
+      );
+    } catch(e) {
+      return false;
+    } finally {
+      await session.endSession();
+    }
 
-    return;
+    return true;
   }
 
 }
@@ -60,25 +72,37 @@ const get = (driver) => get_regular(driver, col(driver));
  * @returns {db_col["remove"]}
  */
 const remove = (driver) => {
-  return async (id) => {
-    const objid = to_objid(id)
-    // const item = await col(driver).findOne(handle_or_id(id));
+  return async (id_or_handle) => {
+    const item = await col(driver).findOne(handle_or_id(id_or_handle));
+    if(!item) return;
+    const objid = to_objid(item.id);
+    const session = driver.mongo_client.startSession();
 
-    ////
-    // STOREFRONTS --> SHIPPING RELATION
-    ////
-    await driver.storefronts._col.updateMany(
-      { '_relations.shipping_methods.ids' : objid },
-      { 
-        $pull: { '_relations.shipping_methods.ids': objid },
-        $unset: { [`_relations.shipping_methods.entries.${objid.toString()}`]: '' },
-      },
-    );
+    try {
+      await session.withTransaction(
+        async () => {
+          ////
+          // STOREFRONTS --> SHIPPING RELATION
+          ////
+          await driver.storefronts._col.updateMany(
+            { '_relations.shipping_methods.ids' : objid },
+            { 
+              $pull: { '_relations.shipping_methods.ids': objid },
+              $unset: { [`_relations.shipping_methods.entries.${objid.toString()}`]: '' },
+            },
+            { session }
+          );
 
-    // DELETE ME
-    const res = await col(driver).findOneAndDelete( 
-      { _id: objid }
-    );
+          // DELETE ME
+          const res = await col(driver).findOneAndDelete( 
+            { _id: objid }, 
+            { session }
+          );
+        }
+      );
+    } finally {
+      await session.endSession();
+    }
 
     return
   }
